+++ packages/docs/src/gatsby-plugin-theme-ui/index.js
@@ -53,8 +53,9 @@ export default {
   fontSizes: [12, 14, 16, 20, 24, 32, 48, 64, 72],
   fontWeights: {
     body: 400,
-    heading: 700,
-    display: 900,
+    heading: 800,
+    bold: 700,
+    display: 800,
   },
   lineHeights: {
     body: 1.5,
@@ -164,6 +165,20 @@ export default {
       border: 0,
       borderBottom: '1px solid',
       borderColor: 'muted',
+    },
+    xray: {
+      '*': {
+        outline: '1px solid rgba(0, 192, 255, .25)',
+      },
+    },
+    navlink: {
+      display: 'inline-block',
+      fontWeight: 'bold',
+      color: 'inherit',
+      textDecoration: 'none',
+      ':hover,:focus': {
+        color: 'primary',
+      },
     },
   },
   prism,